tests/wpt/web-platform-tests/html/browsers/browsing-the-web/overlapping-navigations-and-traversals/tentative/same-document-traversal-same-document-traversal-hashchange.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Same-document traversals during same-document traversals (using fragment navigations)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
  In the spec, all traversals are queued. However, what "back" and "forward"
  mean is computed synchronously. So per spec:

  - back(), back(): go back 1.
  - back(), forward(): go forward 1 (if you're not starting from the end).

  (The spec doesn't care about same or cross-document.)

  This is not how browsers behave:

  - Chrome seems to ignore same-document traversal requests if one is already in flight:

    - back(), back(): go back 1.
    - back(), forward(): go back 1 (regardless of starting at the end or not).

  - Firefox seems to queue up "back" or "forward" commands:

    - back(), back(): go back 1, then go back 1.
    - back(), forward(): go back 1, then go forward 1 (regardless of starting at the end or not).

  We assert the Firefox behavior here for now.
-->

<body>
<script type="module">
import { createIframe, delay, waitForHashchange } from "./resources/helpers.mjs";

promise_test(async t => {
  const iframe = await createIframe(t);
  const baseURL = iframe.contentWindow.location.href;

  // Setup
  iframe.contentWindow.location.hash = "#1";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#2";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#3";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.history.back();
  await waitForHashchange(iframe.contentWindow);
  assert_equals(iframe.contentWindow.location.hash, "#2", "we made our way to #2 for setup");

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.hash, "#2", "must not go back synchronously");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.hash, "#2", "must not go forward synchronously");

  const event1 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event1.oldURL, baseURL + "#2", "oldURL 1");
  assert_equals(event1.newURL, baseURL + "#1", "newURL 1");
  // Cannot test iframe.contentWindow.location.hash since the second history
  // traversal task is racing with the fire an event task, so we don't know
  // which will happen first.

  const event2 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event2.oldURL, baseURL + "#1", "oldURL 2");
  assert_equals(event2.newURL, baseURL + "#2", "newURL 2");
  assert_equals(iframe.contentWindow.location.hash, "#2");
}, "same-document traversals in opposite directions: queued up");

promise_test(async t => {
  const iframe = await createIframe(t);
  const baseURL = iframe.contentWindow.location.href;

  // Setup
  iframe.contentWindow.location.hash = "#1";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#2";
  await waitForHashchange(iframe.contentWindow);

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.hash, "#2", "must not go back synchronously");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.hash, "#2", "must not go forward synchronously");

  const event1 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event1.oldURL, baseURL + "#2", "oldURL 1");
  assert_equals(event1.newURL, baseURL + "#1", "newURL 1");
  // Cannot test iframe.contentWindow.location.hash since the second history
  // traversal task is racing with the fire an event task, so we don't know
  // which will happen first.

  const event2 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event2.oldURL, baseURL + "#1", "oldURL 2");
  assert_equals(event2.newURL, baseURL + "#2", "newURL 2");
  assert_equals(iframe.contentWindow.location.hash, "#2");
}, "same-document traversals in opposite directions, second traversal invalid at queuing time: queued up");

promise_test(async t => {
  const iframe = await createIframe(t);
  const baseURL = iframe.contentWindow.location.href;

  // Setup
  iframe.contentWindow.location.hash = "#1";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#2";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#3";
  await waitForHashchange(iframe.contentWindow);

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.hash, "#3", "must not go back synchronously (1)");

  iframe.contentWindow.history.back();
  assert_equals(iframe.contentWindow.location.hash, "#3", "must not go back synchronously (2)");

  const event1 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event1.oldURL, baseURL + "#3", "oldURL 1");
  assert_equals(event1.newURL, baseURL + "#2", "newURL 1");
  // Cannot test iframe.contentWindow.location.hash since the second history
  // traversal task is racing with the fire an event task, so we don't know
  // which will happen first.

  const event2 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event2.oldURL, baseURL + "#2", "oldURL 2");
  assert_equals(event2.newURL, baseURL + "#1", "newURL 2");
  assert_equals(iframe.contentWindow.location.hash, "#1");
}, "same-document traversals in the same (back) direction: queue up");

promise_test(async t => {
  const iframe = await createIframe(t);
  const baseURL = iframe.contentWindow.location.href;

  // Setup
  iframe.contentWindow.location.hash = "#1";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#2";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.location.hash = "#3";
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.history.back();
  await waitForHashchange(iframe.contentWindow);
  iframe.contentWindow.history.back();
  await waitForHashchange(iframe.contentWindow);
  assert_equals(iframe.contentWindow.location.hash, "#1", "we made our way to #1 for setup");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.hash, "#1", "must not go forward synchronously (1)");

  iframe.contentWindow.history.forward();
  assert_equals(iframe.contentWindow.location.hash, "#1", "must not go forward synchronously (2)");

  const event1 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event1.oldURL, baseURL + "#1", "oldURL 1");
  assert_equals(event1.newURL, baseURL + "#2", "newURL 1");
  // Cannot test iframe.contentWindow.location.hash since the second history
  // traversal task is racing with the fire an event task, so we don't know
  // which will happen first.

  const event2 = await waitForHashchange(iframe.contentWindow);
  assert_equals(event2.oldURL, baseURL + "#2", "oldURL 2");
  assert_equals(event2.newURL, baseURL + "#3", "newURL 2");
  assert_equals(iframe.contentWindow.location.hash, "#3");
}, "same-document traversals in the same (forward) direction: queue up");
</script>
